v2.2 — fix UTF-8 BOM, raise key limits to 64/100, uint64_t changedMask
BOM FIX
  All .ino files written via Set-Content had a UTF-8 BOM prepended,
  causing Arduino IDE compiler errors:
    error: stray '\357' in program
    error: stray '\273' in program
    error: stray '\277' in program
  Fixed by re-writing all affected files with UTF-8-without-BOM:
    examples/Arduino_BT_NoLibrary/Arduino_BT_NoLibrary.ino
    examples/Arduino_RC_Car/Arduino_RC_Car.ino
    examples/ESP32_RC_Car_BT/ESP32_RC_Car_BT.ino
    examples/ESP32_RC_Car_Wifi/ESP32_RC_Car_Wifi.ino
    examples/ESP32_WiFi_NoLibrary/ESP32_WiFi_NoLibrary.ino

KEY LIMIT INCREASE
  Library (RoboLinkParser.h):
    ROBOLINK_MAX_PAIRS on ESP32/ARM:    32 -> 64
    _changedMask:           uint32_t -> uint64_t  (supports up to 64 tracked keys)
  No-library examples:
    Arduino_BT_NoLibrary   MAX_KEYS:    8 -> 16   (AVR RAM ~240 B for store)
    ESP32_WiFi_NoLibrary   MAX_KEYS:   32 -> 100

changedMask uint64_t PROPAGATION
  Updated all uint32_t mask casts to uint64_t in:
    src/RoboLinkParser.h    _changedMask member + changedMask() return type
    src/RoboLinkParser.cpp  _parseLine bitmask ops, changed(), changedMask(), changedCount()
    src/RoboLinkSerial.cpp  update() callback loop
    src/RoboLinkBT.cpp      update() callback loop
    src/RoboLinkWiFi.cpp    update() callback loop

+++ RoboLink/src/RoboLinkBT.cpp
@@ -35,11 +35,11 @@ void RoboLinkBT::update() {
     /* 2. Callbacks AFTER reading — only changed keys */
     if (gotMsg) {
         if (_valueCb) {
-            uint32_t mask = _rx.changedMask();
+            uint64_t mask = _rx.changedMask();
             for (int i = 0; i < _rx.count() && mask; i++) {
-                if (mask & ((uint32_t)1 << i)) {
+                if (mask & ((uint64_t)1 << i)) {
                     _valueCb(_rx.keyAt(i), _rx.valueAt(i));
-                    mask &= ~((uint32_t)1 << i);
+                    mask &= ~((uint64_t)1 << i);
                 }
             }
         }
