+++ RoboLink/src/RoboLinkSerial.cpp
@@ -76,11 +76,11 @@ void RoboLinkSerial::update() {
             /* Only fire for keys that actually changed — reduces
                Serial.print volume and prevents SoftwareSerial
                RX overflow on AVR.  */
-            uint32_t mask = _rx.changedMask();
+            uint64_t mask = _rx.changedMask();
             for (int i = 0; i < _rx.count() && mask; i++) {
-                if (mask & ((uint32_t)1 << i)) {
+                if (mask & ((uint64_t)1 << i)) {
                     _valueCb(_rx.keyAt(i), _rx.valueAt(i));
-                    mask &= ~((uint32_t)1 << i);
+                    mask &= ~((uint64_t)1 << i);
                 }
             }
         }
